Moved Elasticsearch connection management to own module
